tests: Fixup secureboot enrollment test
Also split out Exec, we need it in more tests
